+++ handlers/producer/tx-listener/producer.go
@@ -2,6 +2,7 @@ package txlistener
 
 import (
 	"github.com/Shopify/sarama"
+	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 	"gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/handlers/producer"
 	broker "gitlab.com/ConsenSys/client/fr/core-stack/orchestrate.git/pkg/broker/sarama"
@@ -30,5 +31,24 @@ func PrepareMsg(txctx *engine.TxContext, msg *sarama.ProducerMessage) error {
 
 // Producer creates a producer handler
 func Producer(p sarama.SyncProducer) engine.HandlerFunc {
-	return producer.Producer(p, PrepareMsg)
+	return func(txctx *engine.TxContext) {
+
+		// TODO: Make it possible to filter at the last moment. So that we can produce in multiple topics
+		if ExternalTxDisabled() && (txctx.Envelope.Metadata == nil || txctx.Envelope.Tx == nil) {
+
+			// For robustness make sure that receipt and tx hash are set even though tx-listener already guarantee it
+			if txctx.Envelope.Receipt != nil && txctx.Envelope.Receipt.TxHash != nil {
+				txctx.Logger.WithFields(log.Fields{
+					"tx.hash": txctx.Envelope.Receipt.GetTxHash(),
+				}).Debugf("External tx disabled, skipping transaction with tx hash")
+			}
+
+			// Ignore the following handlers
+			txctx.Abort()
+			return
+		}
+
+		// Else normally call the producer as we would normally do
+		producer.Producer(p, PrepareMsg)(txctx)
+	}
 }